Fix special script

+++ scripts/CI/CI_LETS_DO_THIS_LWS.sh
@@ -23,9 +23,13 @@ mv $LWS.2 $SCRIPT_PATH/LWS_temp.txt
 
 EXPIRED=$MAIN_PATH/expired-domains/LWS_temp-expired.txt
 UNKNOWN=$MAIN_PATH/expired-domains/LWS_temp-unknown.txt
+if [ -f $EXPIRED ]; then
+    comm -2 -3 $SCRIPT_PATH/LWS_temp.txt $EXPIRED >> $MAIN_PATH/LWS/podejrzane_LWS.txt
+    rm -r $SCRIPT_PATH/LWS_temp.txt
+else
+    mv $SCRIPT_PATH/LWS_temp.txt $MAIN_PATH/LWS/podejrzane_LWS.txt
+fi
 
-comm -2 -3 $SCRIPT_PATH/LWS_temp.txt $EXPIRED >> $MAIN_PATH/LWS/podejrzane_LWS.txt
-rm -r $SCRIPT_PATH/LWS_temp.txt
 
 if [ -f "$EXPIRED" ]; then
     rm -r $EXPIRED
@@ -38,6 +42,6 @@ fi
 GIT_SLUG=$(git ls-remote --get-url | sed "s|https://||g" | sed "s|git@||g" | sed "s|:|/|g")
 git config --global user.email "PolishJarvis@int.pl"
 git config --global user.name "PolishJarvis"
-git add --all
+git add $MAIN_PATH/LWS/podejrzane_LWS.txt
 git commit -m "Check [ci skip]"
 git push https://PolishJarvis:${GH_TOKEN}@${GIT_SLUG} HEAD:master > /dev/null 2>&1
